aws/bedrock/AgentCore/pyproject.toml
[project]
name = "agentcore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=0.1.2",
    "bedrock-agentcore-starter-toolkit>=0.1.6",
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=0.3.31",
    "langgraph>=0.6.6",
    "mcp>=1.15.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]
